Refactor AuthenticatedUserService to AccountService
