+++ src/main/java/com/example/echo_api/service/account/AccountService.java
@@ -1,8 +1,8 @@
-package com.example.echo_api.service.user;
+package com.example.echo_api.service.account;
 
 import com.example.echo_api.persistence.model.User;
 
-public interface AuthenticatedUserService {
+public interface AccountService {
 
     public User getAuthenticatedUser();
